Use float for elapsed_time
When accumulating elapsed_time from many moves that take less than 1
second, elapsed_time does not get incremented because (unsigned int)0.9
= 0.

+++ xs/src/libslic3r/GCode.hpp
@@ -82,7 +82,7 @@ class GCode {
     const Layer* layer;
     std::map<const PrintObject*,Point> _seam_position;
     bool first_layer; // this flag triggers first layer speeds
-    unsigned int elapsed_time; // seconds
+    float elapsed_time; // seconds
     double volumetric_speed;
     
     GCode();
